fix the ave_dim_sect in jasmine when truncating the Hilbert space

+++ src/ctqmc/manjushaka/ctqmc_context.f90
@@ -1289,7 +1289,7 @@
      integer  :: nsect_t
 
 ! maximal dimension of sectors after truncation
-     real(dp) :: max_dim_sect_t
+     integer :: max_dim_sect_t
 
 ! averaged dimension of sectors after truncation
      real(dp) :: ave_dim_sect_t
